基础类型方法重构
